build: one cargo workspace for the repo (#133)
* build: one cargo workspace for the repo

* chore: move tests from module -> wasmer_sys to avoid public glob rexport shadowing

* chore: root workspace includes test crate

* chore: changelog

* chore: missed a cargo.lock

* chore: move building test wasms before testing root workspace

* refactor: instead of nesting crates within "test" crate, move all to same level within "test-crates" dir

* chore: rename test_wasm to test_wasm_core to clarify which crates are wasms used for tests vs actual test implementations

* chore: changelog clarity

* fix: path to test crate

* chore: fix bench paths, remove unused script

* fix: windows tests

* wip: debug windows file path

* fix: normalize file path formatting on windows and explain with comment

* chore: fmt

* chore: long arg

* chore: move example wasms into test-crates/wasms dir, use workspace for shared deps

* Revert "chore: move example wasms into test-crates/wasms dir, use workspace for shared deps"

This reverts commit 21d0d6a1648290cc502973a71d9e71e712fe38c7.

* chore: move example wasms into test-crates/wasms dir

* chore: move tests of default behavior of 'wasm_error!()' into common crate where it is defined

* chore: fmt + clippy
