+++ Cargo.toml
@@ -1,5 +1,7 @@
 [workspace]
-members = ["crates/host", "crates/common"]
-exclude = ["test"]
-
+members = ["crates/*", "test-crates/tests", "test-crates/common", "test-crates/wasms/*"]
+default-members = ["crates/*"]
 resolver = "2"
+
+[profile.release.package.tests]
+debug = true
