+++ crates/host/src/module/wasmer_sys.rs
@@ -74,3 +74,48 @@ pub fn get_ios_module_from_file(path: &Path) -> Result<Module, DeserializeError>
     let engine = Engine::headless();
     unsafe { Module::deserialize_from_file(&engine, path) }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::module::{CacheKey, ModuleCache, PlruCache};
+
+    #[test]
+    fn cache_test() {
+        // simple example wasm taken from wasmer docs
+        // https://docs.rs/wasmer/latest/wasmer/struct.Module.html#example
+        let wasm: Vec<u8> = vec![
+            0x00, 0x61, 0x73, 0x6d, 0x01, 0x00, 0x00, 0x00, 0x01, 0x06, 0x01, 0x60, 0x01, 0x7f,
+            0x01, 0x7f, 0x03, 0x02, 0x01, 0x00, 0x07, 0x0b, 0x01, 0x07, 0x61, 0x64, 0x64, 0x5f,
+            0x6f, 0x6e, 0x65, 0x00, 0x00, 0x0a, 0x09, 0x01, 0x07, 0x00, 0x20, 0x00, 0x41, 0x01,
+            0x6a, 0x0b, 0x00, 0x1a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x01, 0x0a, 0x01, 0x00, 0x07,
+            0x61, 0x64, 0x64, 0x5f, 0x6f, 0x6e, 0x65, 0x02, 0x07, 0x01, 0x00, 0x01, 0x00, 0x02,
+            0x70, 0x30,
+        ];
+        let module_cache = ModuleCache::new(None);
+        assert!(module_cache.serialized_module_cache.read().cache.is_empty());
+        assert!(module_cache
+            .deserialized_module_cache
+            .read()
+            .cache
+            .is_empty());
+
+        let key: CacheKey = [0u8; 32];
+        let module = module_cache.get(key, &wasm).unwrap();
+
+        // make sure module has been stored in serialized cache under key
+        {
+            let serialized_cached_module =
+                module_cache.serialized_module_cache.write().get_item(&key);
+            assert!(serialized_cached_module.is_some());
+        }
+        // make sure module has been stored in deserialized cache under key
+        {
+            let deserialized_cached_module = module_cache
+                .deserialized_module_cache
+                .write()
+                .get_item(&key)
+                .unwrap();
+            assert_eq!(*deserialized_cached_module, *module);
+        }
+    }
+}
