+++ scripts/bench-wasmer_sys_dev.sh
@@ -1,10 +1,7 @@
 #! /usr/bin/env bash
-(
- cd test && \
- cargo bench --no-default-features --features wasmer_sys_dev
-)
+
+cargo bench -p tests --no-default-features --features wasmer_sys_dev
 
 # it's possible to flamegraph the benchmarks like this:
 #
-# cd test
 # flamegraph cargo bench --bench bench -- --profile-time 10
