+++ scripts/bench-wasmer_sys_prod.sh
@@ -1,5 +1,3 @@
 #! /usr/bin/env bash
-(
- cd test && \
- cargo bench --no-default-features --features wasmer_sys_prod
-)
+
+cargo bench -p tests --no-default-features --features wasmer_sys_prod
