+++ scripts/test-wasmer_wamr.sh
@@ -5,8 +5,11 @@ export RUST_BACKTRACE=full
 export WASMER_BACKTRACE=1
 
 
-# tests the root workspace that doesn't include any wasm code
+# tests the root workspace
+cargo test --no-default-features --features wasmer_wamr ${1-} -- --nocapture
+
+# tests the root workspace, error_as_host
 cargo test --no-default-features --features error_as_host,wasmer_wamr ${1-} -- --nocapture
 
 # build wasm and run the "full" tests for wasmer_wamr
-cargo test --release --manifest-path test/Cargo.toml --no-default-features --features wasmer_wamr ${1-} -- --nocapture
+cargo test --release -p tests --no-default-features --features wasmer_wamr ${1-} -- --nocapture
