+++ test-crates/tests/Cargo.toml
@@ -1,15 +1,15 @@
 [package]
-name = "test"
+name = "tests"
 version = "0.0.90"
 authors = ["thedavidmeister", "thedavidmeister@gmail.com"]
 edition = "2021"
 
 [dependencies]
-holochain_wasmer_common = { path = "../crates/common" }
-holochain_wasmer_host = { path = "../crates/host", default-features = false, features = ["error_as_host"] }
+holochain_wasmer_common = { path = "../../crates/common" }
+holochain_wasmer_host = { path = "../../crates/host", default-features = false, features = ["error_as_host"] }
 holochain_serialized_bytes = "=0.0.55"
 serde = "1"
-test_common = { path = "./common", default-features = false }
+test_common = { path = "../common", default-features = false }
 criterion = { version = "0.5", features = ["html_reports"] }
 rand = "0.8"
 serde_bytes = "0.11"
@@ -33,9 +33,6 @@ harness = false
 name = "test"
 crate-type = ["cdylib", "rlib"]
 path = "src/test.rs"
-
-[profile.release]
-debug = true
 
 [features]
 debug_memory = ["holochain_wasmer_host/debug_memory"]
