+++ test-crates/wasms/wasm_core/Cargo.toml
@@ -1,16 +1,16 @@
 [package]
-name = "test_wasm"
+name = "test_wasm_core"
 version = "0.0.90"
 authors = ["thedavidmeister", "thedavidmeister@gmail.com"]
 edition = "2021"
 
 [lib]
-name = "test_wasm"
+name = "test_wasm_core"
 crate-type = ["cdylib", "rlib"]
 path = "src/wasm.rs"
 
 [dependencies]
-holochain_wasmer_guest = { path = "../../crates/guest" }
-test_common = { path = "../common", default-features = false }
+holochain_wasmer_guest = { path = "../../../crates/guest" }
+test_common = { path = "../../common", default-features = false }
 holochain_serialized_bytes = "=0.0.55"
 serde = "1"
