+++ test-crates/wasms/wasm_empty/src/wasm.rs
