+++ test-crates/wasms/wasm_io/Cargo.toml
@@ -10,8 +10,8 @@ crate-type = ["cdylib", "rlib"]
 path = "src/wasm.rs"
 
 [dependencies]
-holochain_wasmer_guest = { path = "../../crates/guest" }
-test_common = { path = "../common", default-features = false}
+holochain_wasmer_guest = { path = "../../../crates/guest" }
+test_common = { path = "../../common", default-features = false}
 holochain_serialized_bytes = "=0.0.55"
 paste = "0.1"
 serde = "1"
